stylecheck/requirements.txt
flask==2.3.3
werkzeug==2.3.7
mistralai
anthropic
openai
google-generativeai
python-dotenv
transformers
torch
textblob
nltk
numpy
pandas
sentencepiece
matplotlib
seaborn
textstat
scikit-learn